FIXED task 1 and 2

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -8,10 +8,11 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i = 0;
+	size_t i = 0;
+
 	if (size > 0 && action)
 	{
-		for (i = 0; i < (int)size; i++)
+		for (i = 0; i < size; i++)
 		{
 			(*action)(array[i]);
 		}
